Move Network config out of web3Context
