+++ constants/onboard.ts
@@ -1,6 +1,6 @@
 import { NETWORKS } from '@tracer-protocol/pools-js';
 import { Initialization } from '@tracer-protocol/onboard/dist/src/interfaces';
-import { DEFAULT_NETWORK, networkConfig } from '@context/Web3Context/Web3Context.Config';
+import { DEFAULT_NETWORK, networkConfig } from '~/constants/networks';
 
 export const onboardConfig: Initialization = {
     networkId: parseInt(DEFAULT_NETWORK),
